+++ src/cs2/entity/planted_c4.rs
@@ -2,13 +2,13 @@ use glam::Vec3;
 
 use crate::cs2::{CS2, entity::player::Player};
 
-#[derive(Debug, Clone, Copy, PartialEq)]
+#[derive(Clone, Copy, PartialEq)]
 pub struct PlantedC4 {
-    handle: u64,
+    handle: usize,
 }
 
 impl PlantedC4 {
-    pub fn new(handle: u64) -> Self {
+    pub fn new(handle: usize) -> Self {
         Self { handle }
     }
