+++ src/cs2/entity/weapon.rs
@@ -82,12 +82,12 @@ impl Weapon {
         if handle == 0 {
             return None;
         }
-        let index = handle as u64 & 0xFFF;
+        let index = handle as usize & 0xFFF;
         let entity = Player::get_client_entity(cs2, index)?;
         Some(Self::from_entity(entity, cs2))
     }
 
-    pub fn from_entity(entity: u64, cs2: &CS2) -> Self {
+    pub fn from_entity(entity: usize, cs2: &CS2) -> Self {
         let weapon_index: u16 = cs2.process.read(
             entity
                 + cs2.offsets.weapon.attribute_manager
@@ -97,11 +97,11 @@ impl Weapon {
         Self::from_index(weapon_index)
     }
 
-    pub fn clip_ammo(entity: u64, cs2: &CS2) -> i32 {
+    pub fn clip_ammo(entity: usize, cs2: &CS2) -> i32 {
         cs2.process.read(entity + cs2.offsets.weapon.clip_primary)
     }
 
-    pub fn reserve_ammo(entity: u64, cs2: &CS2) -> i32 {
+    pub fn reserve_ammo(entity: usize, cs2: &CS2) -> i32 {
         cs2.process.read(entity + cs2.offsets.weapon.reserve_ammo)
     }
